+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # Brassica changelog
 
+## Unreleased changes
+
+- Allow grapheme to begin with star
+
 ## v0.1.1
 
 - Rewrote executables with a client/server architecture for better Windows support.