+++ ChangeLog.md
@@ -6,6 +6,7 @@
 - Allow backreferences to occur in the environment
 - Allow user to choose separator used between multiple results (previously a space)
 - Internal refactor: category expansion is now separate from parsing
+- Add `--version` command-line option
 
 ## v0.1.1