+++ ChangeLog.md
@@ -10,6 +10,7 @@
 - Store `MultiZipper` data in a `Vector` rather than a linked list (for performance)
 - Bugfix: subtraction now removes all subtracted graphemes
 - Store paradigm builder output in a tree data structure, allowing a more compact output format
+- Documented abstract features in paradigm builder (previously present but undocumented)
 
 ## v0.1.1